moved `expanderButton.tsx` into `components` dir as it was mistakenly too high up in the directory structure
